refactor: renamed ScheduleTime into TrainStationTime
