+++ src/test/scala/ulisse/entities/timetable/TimetableTest.scala
@@ -7,7 +7,7 @@ import ulisse.TestUtility.in
 import ulisse.entities.Routes.TypeRoute.AV
 import ulisse.entities.Routes.TypeRoute
 import ulisse.entities.station.Station
-import ulisse.entities.timetable.ScheduleTime.{AutoScheduleTime, EndScheduleTime, StartScheduleTime}
+import ulisse.entities.timetable.TrainStationTime.{AutoScheduleTime, EndScheduleTime, StartScheduleTime}
 import ulisse.entities.timetable.Timetables.{toWaitTime, PartialTimetable, RailInfo, TrainTimetable}
 import ulisse.utils.Times.FluentDeclaration.h
 import ulisse.utils.Times.{ClockTime, ClockTimeErrors}
